workers/requirements.txt
# Worker Requirements for Bank of Canada MLOps Platform

# Data processing
pandas==2.1.4
numpy==1.25.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Caching and message queue
redis==5.0.1
celery==5.3.4

# HTTP client for API calls
requests==2.31.0

# Environment configuration
python-dotenv==1.0.0

# Logging
structlog==23.2.0

# Data validation
cerberus==1.3.5
marshmallow==3.20.2

# ML libraries (for data quality checks)
scikit-learn==1.3.2

# Utilities
click==8.1.7
